removed noexcept from defaulted special member functions
Letting the compiler figure it out is less confusing.

Also, MSVC++ won't compile build_scope_from_bindings() with Scope's defaulted move constructor explicitly noexcept, due to the non-noexcept-movable subobject.

I'm not confident that MSVC++ is wrong here, though Clang and GCC are fine with it. See this example:

https://godbolt.org/z/Ybg-nt

This relates, at least vaguely, to http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#1402.

I'm not entirely sure how C++17 is supposed to treat an explicitly noexcept'ed defaulted move constructor on a class with a member with a non-noexcept move constructor.

Perhaps the strangest part of all this is what appears to have triggered it: unordered_map in the MS STL is not noexcept move constructible (for any template arguments):

https://github.com/microsoft/STL/blob/master/stl/inc/unordered_map#L200