properly diagnose malformed terms
I'm using .to_u16? again. Capturing a variable whose type is
restricted due to nil tracing (a form of flow-based typing)
takes it as the original unrestricted type, but capturing a copy
works fine.

Presumably it works this way because, in general, a captured
variable of a sum type could be written to again, with an object of
a different runtime type. So the alternatives are to permit
something that violates type safety, or to adopt considerably more
complex rules to separate the safe cases from the unsafe ones.