named for part A, separate toposort
It will just be for part A, for starters.

Various separately implemented general topological sorting
algorithms could be used. I've changed -kahn to -ts-kahn to denote
that this uses a separately implemented general toposort (-ts)
which is implemented via Kahn's algorithm (-kahn).

That way, if I make a version that uses Kahn's algorithm
specifically for this restricted case, it can be named with -kahn
without -ts.