even this is technically not brute force
A brute-force solution would be one where every path is tried *separately*.

Here, the paths are tried with overlapping *prefixes*.

Of course, this is still horrendously slow because the paths are not tried with
overlapping *suffixes*, which is what a memoization (or tabulation) solution
would do.

In general, naive backtracking algorithms are not technically brute force.