Journey: used hand-rolled Frame class for Ruby version to use fewer fibers
This passed.

With restrictive ulimits, storing a stack of enumerators from .each
can be a problem because each enumerator uses a separate fiber.

I'm not totally sure why ulimits that are high enough for most code
to run, as on HackerRank, would cause this. The whole point of
fibers is to be lightweight.

Perhaps I am misunderstanding the problem. Implementing the same
logic with my own class in place of an array enumerator works,
though.

No corresponding problem appears to arise with the Crystal version
(though I can't directly test that on HackerRank, which doesn't
support Crystal). I'm not sure if Crystal creates fibers to
implement .each without a block. (Of course neither language's
implementations use a fiber when a block is passed.)

Also possibly relevant is that .each in Crystal (without a block)
returns a non-throwing iterator (unless the underlying collection
implements throwing logic). When .next is called on that iterator
but there are no more items, it returns an instance of the class
Iterator::Stop.

This gives rise to the semipredicate problem in the unusual case
that an object of type Iterator::Stop is actually stored in the
collection (getting to it, with non-inlined iteration, ends the
iteration). But in general it's quite a bit nicer if one is to
work with it directly.

Ruby's behavior is similar to Python's -- they both raise a
StopIteration exception when iteration is done. But in Python it
is idiomatic to use exceptions for control flow. In most languages,
incluidng Ruby, it is not, so code to use enumerators directly in
Ruby is unusual and off-putting.