added top comment, submitted, passed
It didn't really need major changes, I don't think. (I hope not,
since I didn't make any major changes.)

I had thought, based on the output of perlcritic, that it might be
a problem that I was using substr as an lvalue.

But I think that is much clearer than calling the substr with four
arguments and discarding the result, and it performs as well in all
remotely recent versions of Perl, I believe:

https://stackoverflow.com/a/61678561

I do use the four-argument way when I want the return value.