scratchwork sort of showing what is happening
Values added to heaps, starting with the value 4 in the example
shown in the scratchwork added here, just disappear, being replaced
by other values.

I think the problem may have to do with an assignment of the form
(x, y) = (y, x). Is mcs implementing an old C# dialect that doesn't
have ValueTuple support and, if so, is this involving the comma
operator?

I would expect that sort of thing, in C#, to give a compile error,
or at least a warning. But perhaps not.