tree-based implementation of Median Updates
The heap-based implementation worked, and the optimized version,
median-updates-fast.cs, passed all the test cases. (The unoptimized
version, median-updates.cs, passed all but one test case, which
timed out. Running that test locally did produce the correct
result.)

But this approach is simpler, and given the complexity of the heap-
based way (with two layers of hash lookups), this should be faster.

The usual tree-based approaches to Median Updates uses one or two
multisets. Some languages' standard libraries, like that of C++,
have such a data structure. Or one may implement it from scratch.
(Or, outside the setting of "competitive programming," use an
implementation from a high-quality third-party library.)

Equipped with tree-based multisets, one may use a two-tree approach
analogous to the two-heap approach typically used to solve the
easier problem where values are only ever added, never removed.

Or, if one can step back and forth in the tree, one can use a
single tree and keep track of the medial node(s) in the order they
would be reached in inorder traversal.

Or, if one has an order statistic tree, one may use that.

But there are two other ways to implement a tree-based multiset.
Many frameworks provide a set but not a bag/multiset. .NET/Mono is
such a framework. (There is a ConcurrentBag class, but it is ill
suited for this use.)

One is to attach a unique identifier to each element in the tree,
to break ties and allow duplciates to effectively be stored. I've
done that before (though I haven't redone it in this repository).

The other is to use the tree set to store the frequency of each
element. The cardinality should also be manually tracked, so it can
be looked up in O(1) -- since the underlying tree set type tracks
how many unique elements there are, rather than how many elements.

I'm not sure I've actually implemented it that way before, even
though it is conceptually (and, I believe, practically) quite
simple. That is the approach I've taken here.

Currently, this is only very minimally tested.