started on C++ "Swiss Tables" version
This uses absl::flat_hash_map and absl::flat_hash_set instead of
std::unordered_map and std::unordered_set.

I'm having trouble getting this to link. It uses Abseil in way that
is not just header-only, and I don't know which binaires to link
against. Both with CMake and when compiling manually with
g++/clang++, I don't know how to make this link.