Snakes and Ladders: The Quickest Way Up, initial version
In spite of having solved this in the past on the first try with no
trouble whatsoever, I seem to be making some strange mistakes.

This version times out on most test cases because I failed to keep
a visitation list. But I believe there are other serious bugs as
well.