improved regex readability for part A
This regex is not actually equivalent, but it works just as well,
and in the same way. It matches word characters instead of
characters other than specific punctuation. Since all the non "["
and "]" characters in the input are word characters, this works as
well and is far more readable (due to "\w").

The corresponding readability approach is already taken in my
solution to part B (day07b.pl).