AOC2015 9a, but my answer is wrong apparently
Suspiciously, it's always picking the first edge specified in the
input. Since vertex numbers are chosen in increasing order of
appearance in the input, this means it's picking the
lexicographically earliest tour (by vertex number, not by city
string) instead of the lowest-cost one.

So it looks like there's a pretty severe bug somewhere. Maybe one
of the data structures is not being properly populated?