Advent of Code 2020, day 23, part A, via linked list
See previous commit for most changes; this commit is just
refactoring and cruft removal.

This technique is overkill for part A but I intend to use it for
part B, so I figured I'd do it for part A first (since the extra
work is small, but more importantly so I have a convenient small
list input size for debugging).