Find the Running Median, in Ruby - PriorityQueue bugfix
This bug was in the code that raises an exception when top or pop
is called on an empty priority queue. Since this is a logic error,
which did not appear in the code, that code path was not exercised,
so this bug in my PriorityQueue class didn't cause the solution to
fail any HackerRank test cases (indeed, the solution overall should
not have failed on any input as a result of this, but this is still
a bug).

The bug was left over from the Crystal version, which I wrote and
then manually translated to Ruby. In Crystal, a[i] raises an
exception when there is no such element -- to get nil instead one
uses a[i]? -- but in Ruby a[i] does what a[i]? does in Crystal.

The fixed solution still passes.