just sort to heapify
This makes the heapifying step asymptotically slower, but the
constants are what matter, and the algorithm as a whole is not
even asymptotically slower. I think Ruby's built-in sort is, in
practice, faster than the repeated sifting-down heapifying
algorithm, until very large sizes are reached.

However, the test cases on LeetCode have not managed to distinguish
the two by performance. So although this still passes and it might
be faster for the input sizes in the test cases, I don't know that.

The usual heapifying algorithm is sufficiently simple that there is
little to lose by keeping it, and since I don't know this is
working any better, I'll probably revert this change.