fixed bugs other than the conceptually central one
This had two *serious* bugs:

(1) Considering depth to increase with every source vertex
    processed rather than with every layer. I don't think I've ever
    made that mistake before. (This is reflected in the major loop
    that needs to be nested not being nested.)

(2) Relaxing destination vertex costs using the best cost to the
    source on any path, rather than the cost on the current path.
    This wouldn't make a difference in Dijkstra's algorith (since
    they will always be the same), but in BFS it does make a
    difference when maximum depth is limited, because one can
    relax to a cost achievable only with greater depth than the
    currently recorded depth.

This commit fixes (1) but does not fix (2), so this is still not
correct. In addition, this solution would benefit from being mostly
rewritten, since one *can* use the same element type both in rows
of the adacency list and in the queue, but this is very confusing.
(The best is really to use a tuple type in a language with
structured bindings or equivalent, e.g., C++, C#, Python, Ruby, but
this is a Java solution and that's not available, so members are
being accessed by the name they are given in a user-defined type,
so the name should actually make sense.