Cheapest Flights Within K Stops - correct solution
This passes. I've rewritten (the most important parts of) this
solution, fixing bug (2) described in e143c37 (se also the comments
in the code introduced in this commit, explaining why it must be
done this way).

I'm also using two different data types for adjacency-list row
elements and queue elements (since they are not conceptually the
same). More info about that issue is also in the e143c37 commit
message.