in Python, has bug
The bug appears to be due to how I lazily (and not in the good way)
wrote the loop that adds the edges to the graph in the top-level
function and used, as one of the loop variables, the same variable
name as in one of the function parameters.

(Even if somehow this is not the cause of the bug--which is very
unlikely--it is in and of itself a bug and must be fixed. The
variable name could be changed, but it is better to extract the
code that builds the graph to a separate function... as I nearly
always do, but for whatever reason laxly didn't bother with this
time.)