alternative naive recursive implementation
This optimizes for the case where the subtree actually consists of
the same node objects (in memory) as a part of the tree being
searched.

This passes, but it is no faster than the constexpr version that
doesn't have this optimization, and possibly slower. That makes
sense, since the trees are unlikely to be created in sucha a way as
to actually coincide in memory. (But I wanted to see.)