LeetCode: Serialize and Deserialize Binary Tree, all-iterative
This commit fixes the bug from the preceding commit where two
null tokens (".") instead of one were emitted for left nulls of
nodes with right children.

It passes.

This is an all-iterative version (i.e., the "norec" version, where
not only is deserialization iterative, as in the original, but
serialization as well).