LeetCode: Flatten Binary Tree to Linked List
Recursive and iterative solutions, both passed.

These solutions use auxiliary space proportional to the height of
the tree. A method that uses constant auxiliary space should be
possible, and is suggested as a "Follow up" in the problem
description. I have not written such a solution (yet). It might be
possible along the lines of a threaded (Morris-style) traversal, if
that can be adapted for preorder (rather than inorder) traversal.
Though even if so, I'm not sure if that would be the simplest way.