LeetCode: Couples Holding Hands - wrong attempt
This is not quite as straightforward a permutation problem as I was
(somehow, who knows how) thinking it was. The ways I was trying, in
the two wrong files committed here, did not always pick a workable
representative for each bucket, since this has to be done in an
order that is not initially obvious.

In effect, the problem is that the cycles are undirected, rather
than directed. A combinatoric technique that follows the cycles,
picking the correct direction as it goes, should work, but I find
the prospect of implementing that confusing.

But all components are cycles in this problem, since each vertex
has degree 2. (Singleton components are arguably two cycles, since
they have the edge from the vertex to itself, and the edge from the
vertex "back to" itself, but this is not a problem--the number of
swaps needed is the number of vertices minus the number of cycles,
and singleton cycles removed their vertex from being needed in any
swap, so they behave exactly like single cycles.) So any method of
counting components will work.

I'm committing this wrong code for, uh, posterity, but I plan to
immediately revert this commit and commit undirected-graph component
counting solutions instead.