LeetCode: Couples Holding Hands
Recursive DFS, iterative DFS, BFS, and union-find solutions,
counting components, since each component is a cycle (due to each
vertex being of degree 2) and the minimum number of swaps needed is
the number of vertices minus the number of cycles.

(Couples that are already seated together contribute parallel loops,
but counting this as one cycle rather than two is what's needed for
this problem, so this does not need to be handled separately.)

All passed.

(A correct swapping combinatoric solution should also be possbile,
but I haven't done that.)