slightly simplified recursive logic
By using the safe navigation operator, as I did in the iterative
doubly linked list version. This passes. But it seems slower than
the other way, and unlike in the iterative version (where extra
state is not stored), there is no major advantage to doing it this
way in the recursive implementations. So I'll likely revert this
change.