LeetCode: Course Schedule II - recursive DFS, deconstructing
In Ruby. Recursive DFS implementation. Deconstructs the graph,
removing edges as they are traversed. This has no significant
advantages over just looping through the edges (since a vistation
list must be maintained anyway). However, the intuitive way to
adapt this version of the algorithm to be implemented iteratively
(not yet written, not in this commit) is easier and simpler (and in
some languages tends to run faster) than an iterative
implementation of the non-deconstructing way.

Passed.