Course Schedule II - iterative DFS, deconstructing
In Ruby. Iterative DFS implementation. Deconstructs the graph,
removing edges as they are traversed. Compared to an iterative
non-deconstructing DFS implementation, this is easier and simpler
(and seems to be faster, in Ruby anyway), because there is no need
to keep track of current positions in rows of the adjacency list,
since neighbors (destination vertices) are removed from a row when
they are used.

Passed.