LeetCode: Shortest Path with Alternating Colors - old wrong attempt
I wrote this on 18 October 2019, so it would normally go in the
old-practice repository rather than here. But I never got around to
writing a correct solution at that time. Since my correct solution
(which I'll commit here soon) is neither based on nor otherwise
informed by or related to this wrong attempt, I figured I'd put
the old wrong attempt here for posterity, rather than putting new
solutions (that aren't dervied from old ones) in old-practice or
putting a wrong solution but no correct solutions to the same
problem in old-practice.

The main problem with this wrong solution is that it doesn't keep
track of visitation properly. I am not very inclined to try to fix
this up: the minimal fix would still, I think, be a major change,
and the other approach of using a doubled bipartite graph is much
simpler, due to its modularity (building the graph and implementing
BFS can be entirely independent, with the BFS implementation not
knowing about coloring).