LeetCode: 3Sum - TLE
Although correct, this fails with time limit exceeded. The problem
is that duplicates are produced and then removed, so with many
dupliated inputs, a lot of work is done, only for most of the
results to then be dropped as duplicates. A better approach is to
avoid making the duplicates in the first place.