fixed distinct-elements bug; passes
If a node's value is half the target value, then it matters whether
or not the addends have to be the values of distinct nodes. This
matters whether or not the tree may contain repeated values. There
are test cases for which this matters.

The problem statement is ambiguous with respect to whether, when it
says the tree is a BST, this includes the requirement that all
node values be distinct. This is a separate matter and I don't know
if there are test cases that depend on it, but this solution should
work either way, since the upmost node with the value being searched
for is found. If this value is half the target value but there is
another node in the tree with the same value, then when each_node
recurses to that lower node and searches the BST from its root, the
upper node will be found and match it.