Path With Minimum Effort, bisecting over BFS, fixed bug
The problem was that I was bisecting on a sorted array of edge
weights to find the minimum edge weight for which all edges
traversed would be no more than this weight, but a 1x1 grid graph
has no edges (and the solution is 0). I've handled this separately.

Now that the bug is fixed, this passes.