+ insertion sort, binary insertion sort
Not submitted. I expect these to be too slow to pass on LeetCode,
coded as they are in Python (especially with CPython as the
interpreter). Test cases are up to 50,000 elements, which a binary
insertion sort could probably finish in time in C or C++ if the
shifting were written in such a way as to be vectorizable. But not
this way.