some optimizations, still too slow
This eliminates words with letters not appearing anywhere on the
board, as in 102b015. It also removes words from the trie, but
leaves all chains intact, when they are found, to avoid adding
different board occurrences of the same words to the matches
collection and having to deduplicate them.

I haven't submitted this but it is clearly too slow. It's TLE
about half the time when testing on an input where a word can be
matched in many different ways on the board. I think I need also
to prune chains that no longer reach any (not yet found) words.