Word Search II, simplified considerably, still passes
There is no need to keep track of the subtree populations, because
the only important distinction is between those of zero and nonzero
populations, and when retreating from a subtree with zero words
stored, everything has already been pruned away, so there are no
local children.

This is faster than the more complicated solution.