simplified even more, still passes
I've just done a single check if the child subtree is empty, at the
very end of the recursive helper function, instead of every time it
may become empty.

This is at least as fast, perhaps faster. (In some other languages
it might be faster to do more of the checks, I think.)