renamed license note
The "license note" is a note on fair use, and how I have decided
when to included or omit (1) scaffolding code written by problem
setters and (2) small test cases that accompany problems and are
effectively part of their description. But it is not actually a
qualification on the license under which this project as a whole is
offered. (If it were, then nearly any nontrivial project would have
to have claims about its licensing accompanied by disclaimers.)

But the impression that this project is in some sense not really
0BSD-licensed was wrongly given, because the file was named in such
a way as to be consulted by https://licensee.github.io/licensee/
(which GitHub uses for license detection) as specifying -- or in
this case as rendering ambiguous -- what license the project as a
whole is offered under. Renaming it makes clearer to humans what
it is about -- fair use -- while refraining from misleading GitHub.