instrumented to find bug
The bug is that words cannot be shortened by removing a character,
because removals at different positions can produce the same
result. For example, removing the third character of "tail" and the
fourth character of "talk" both produce "tal", but there should be
no edge in the graph between "tail" and "talk".