in AoC 2021 day 15, fixed bug where intended optimization did nothing
This commit also enabled some more runtime type-checking in part B
(in part A it was already enabled everywhere). This was not really
faciltiated by fixing the optimization. I had hoped to be able to
turn it on everywhere, but instead I left it off of all the simpler
Grid methods that are called repeatedly in loops.