LeetCode: Detonate the Maximum Bombs (too slow)
This naive solutions appears correct, but it is too slow. It fails
with time limit exceeded.

A more efficient approach would find the strongly connected
components and build the metagraph in which they are each contracted
into a single vertex labeled by the population of vertices that
formed it. This all takes only linear time, and this metagraph is
a DAG. Then the total population reachable from each metagraph
vertex can also be computed in linear time (in the size of the
metagraph, which is no greater than the original graph and often
much smaller), in a way resembling topological sort (analogous to
finding shortest paths in a DAG).