fixed bug where vertices were not reachable from themselves
...and thus not from any other vertices either (transitively).

Note that this is not the only bug, and it doesn't seem to be the
bug I actually observed in testing. There may be numerous other
bugs, but the other one I'm aware of is that my implementation of
Kosaraju's algorithm is apparently wrong and always returns empty
groups of vertices (when it should instead *never* return empty
groups of vertices, since a strongly connected component has at
least 1 vertex).