Fix one of many misnamings of stack-based search
See:
https://11011110.github.io/blog/2013/12/17/stack-based-graph-traversal.html

There are a number of others in this repository (but search and
replace would not work, since sometimes I do *actually* implement
what really is an iterative DFS, such as with a stack of frames or
a stack of iterators). Also, in trees, including binary trees
(there are many binary tree problems in this repo), stack-based
search really is an iterative DFS. So the full problem of such
misnamings in this repository will require careful examination to
fix.

I happened to be looking at this code for an unrelated reason, so I
figured I'd fix the misnaming here while I was at it.