Start on LeetCode: Sliding Window Maximum
I misunderstood the problem. I thought it was asking for a sliding
window that contains a maximal element. The wording of the problem
is (as others have pointed out) somewhat ambiguous. However, the
example in the problem statement does definitiely show that this is
not the intended interpretation. An even better reason I should've
recognized this was not intended interpretation is that the kind of
sophisticated approach the problem appears to require, and that I
implemented here, is way more complicated (and less computationally
effiicent, including asymptotically) than it needs to be if the
problem is to be interpreted that way.

Nonetheless I am committing this to have a record of my process
and, more so, because it should be easy to adapt this to solve the
problem that was actually intended (since the general approach is
reasonable for that problem, even as it is unreasonable for the
problem I wrongly thought was described).